Time: 8 ms (53.68%), Space: 16.6 MB (72.35%) - LeetHub

+++ 1022-sum-of-root-to-leaf-binary-numbers/1022-sum-of-root-to-leaf-binary-numbers.cpp
@@ -11,7 +11,7 @@
  */
 class Solution {
 public:
-    int find_sum(TreeNode* root,long long int num)
+    int find_sum(TreeNode* root,int num)
     {
         
         if(root==NULL)
